FIX lti1p3 Entrypoint::Launch, Authen::LTI1p3
Adapt them to Mojolicious.

The Launch::_verifyMessage() calls Authen::LTI1p3 and needs a course
specific CE and DB object. When Launch::run() calls _verifyMessage(), it
was before we even know what course we're in, so I had to move where
it's called to after we've got a course specific CE and DB.
Authen::LTI1p3 requires WeBWorK::Authz to be instantiated, so I've also
created that.

I've updated the section pertaining to launching directly into an
assignment set, but I'm not sure if it's going to work. Need to test.

Main issue in Authen::LTI1p3 is that the JSON::Validator library has
deprecated how we used to use it, so we now have to load every single
schema into the validator manually first before calling validate(). It
used to just pull in dependent schemas automatically.