FIX lti1p3 launch directly into a hw set
Broken because it's not looking in the right place for the set name.

Note that the way we're telling instructors to set up direct hw set
links in Canvas results in us attaching "?set=HW_SET_NAME" to the target
link URL.

So we have to look for the set name in the Target Link URI. Now, due to
the change where the redirect URL is no longer identical to the target
link URL, we have to grab the Target Link URI from the id_token claims.
Previously we just sent the LTI redirect request to the target link URL
which also makes the set name available as a query param, saving us from
having to grab it from the claims.