EDIT student progress page now has # of attempts
Only for quizzes/tests. This restores a customization we had that was
partially wiped in the merge. The calculation part remains in Utils, we
just needed to add back the frontend part.

Instructors can allow multiple tries at a test so students can fix their
mistakes. There is now a # of attempts in the Student Progress page for
that test that'll show how many submits a student used for each version
of the test they took.

Also deleted the LTI1p3Status page that nothing appears to be using.

Launching into quizzes was broken due to ProblemProcessing using the old
namespace for LTI1p3.