FIX launch into quiz, launch auto course create
Just noticed that the rename of 'quiz' into 'test' extends to urls. So
when launching directly into a quiz, the URL used to be "/quiz_mode/"
and is now "/test_mode/".

The messages from launching into a not yet opened test or course create
were getting lost since they hadn't been migrated during the upgrade.
I've switched them to using the Mojolicious flash system, which
surprisingly Webwork itself isn't using.

Course creation on LTI launch wasn't working because it was trying to
sign the user in before the course tables have been created. In order to
make this work, I had to separate out verification of the id_token from
the normal authen flow. This lets us verify the id_token separately from
signing the user in. So Launch.pm's flow is now first verify the
id_token, then create/update the course, then call the normal authen
verify to sign the user in.