FIX user setting 'use equation editor' not saving
Using the MathView editor, when a user tries to set the "Use Equation
Editor?" option to no via "User Settings", it doesn't save the setting.
It just reloads the page with the option back to yes.

It seems the "useMathView" option was accidentally removed from
Options.pm, so I've just restored it.

This is a redo of commit: 9bf034c041e199979cd25fc54cc4f58f605ba9d8