EDIT gateway quiz restore autosave
Restore the old quiz autosave javascript that used to be GatewayQuiz.pm.
This is now in gateway.js. Also tried adding some better error handling
by parsing the autosave return to see if it's telling the user to login
again. This will hopefully reduce the amount of people who lose their
quiz answers because they signed into another device/browser and
invalidated their session.

When external login is on, Login.pm displays an error message directing
users to go back to the external system for login. The messages requires
some LTI related params to be set in localOverrides.conf so a link to
the external LMS can be generated, I've added these params to
localOverrides.conf.dist.

There is still an issue with the external login state not being
persistent, which should be solved in the next commit.