FIX minion job queue can't start
Default settings uses sqlite for the Minion job queue. Turns out sqlite
wasn't installed in the container and installing sqlite3 allowed the
Minion queue to start.

However, sqlite isn't a good fit for our deployment. Mainly because we
would have to setup data persistence for the new database, otherwise the
queue might be lost when the container restarts with an updated image.
We also have several nodes, so it would be best if the queue can be
shared among multiple workers. For those reasons, I decided to try out
the mysql backend for Minion, so we can reuse the same database that
webwork runs on.

Surprisingly, the mysql Minion backend setup was pretty straightforward.
After installing the package and configuring the db authentication
credentials, it just started without issues. It's able to create the
Minion tables automatically.

I've also moved the Minion worker into its own service container, this
is to maintain the Docker best practice of having only 1 process running
per container. The Minion worker service is called the worker_mojo
to distinguish itself from our existing worker. The existing worker was
renamed to "worker_lti1p3" since it only executes LTI1p3 jobs.