EDIT GatewayQuiz.pm save all answers
Webwork 2.15 saved all answers, even answers sent via preview. This was
very helpful for troubleshooting and academic integrity investigations.
Webwork 2.16 changed it so that only full submit answers are saved.

I've swapped it back to save all answers. Previews or autosave answers
are marked separately. An additional column indicating "preview" or
"autosave" was added to the answer log output. For the past_answer
table, "preview" or "autosave" was added as an answer_string prefix.

This is a redo of commit 5f435238ae03ba53a6ca99b4c95656f1fc870ca5