ADD Grade Saved Drafts feature in Student Progress
At the bottom of the Student Progress page, if you're looking at
progress in a timed test/quiz, you'll see a "Grade Saved Drafts" button.
This is so that instructors can mass submit tests that students started
but didn't get to submit for whatever reason.

This is a redo of commit bda7c28d717718c769f152c24cb7b581ac126d24

I tried to improve on the previous implementation by splitting it up
into more chunks and using PG to calculate the score like the original
GatewayQuiz. As noted in a comment in the original commit, it's probably
better to grade the answer saved in *_problem_user, so I also tried to
do that. The original implementation is such a long deeply nested
function though, that I ended up feeling the improvements are minimal.
At least we're using PG to calculate the score now, which decreases the
differences we have from the stock GatewayQuiz grading.

_getProblemHTML() is copied from GatewayQuiz.pm but with some values
switched to hardcoded values since we don't need to use the actual
rendered result, we only want the score.

One thing I did to simplify the implementation is assume that we're not
doing late sumbit calculations. Since this is timed test/quiz, students
wouldn't be able to save progress once their time is up anyway.

I had to make initialize() async since _getProblemHTML() needs to be
called async.