EDIT use envvar for debug, install vim
Allow using the DEBUG env var to override debug settings. The old debug
setting was completely undone in the merge, and the new debug seems to
default to stdout. So the old debug.log is no longer present. Maybe
we'll want to output to both stdout and debug.log in the future, but for
now, we can try just the stdout.

Might as well build in vim into the image, so I don't need to install it
manually when debugging.