FIX lti1p3 redirect uri using wrong scheme
In our kubernetes cluster, https is setup at the ingress, so we don't
need to explicitly setup Webwork's hypnotoad server for https. This
results in Webwork generating http urls for the redirect_uri. Canvas
expects a https url and rejects the http redirect_uri. So I've switched
to using the WEBWORK_ROOT_URL env var as the base of the redirect_uri.
This should properly give us the right http or https redirect_uri.