FIX lti1p3 launch broken during dst jump forward
During daylight savings time, on the time day before we jump forward,
there is an hour in which lti launch will be broken as all nonces will
be treated as expired. This is due to the TIMESTAMP column doing
automatic timezone conversion. So even though we're storing UTC
timestamps, this conversion causes us to fall into an edge case where
it's impossible to convert the UTC time to a DST timezone accurately.

The solution is to change the TIMESTAMP column to DATETIME which doesn't
do any time zone conversion.

Note that for existing tables, they need to be manually migrated with:

ALTER TABLE lti_nonces MODIFY expires_at DATETIME;