EDIT set Mojo secret from WEBWORK_SECRET env var
Mojolicious uses a secret passphrase for security features such as
signed cookies. Since this should be uniquely configurable for each
environment, I've customized it so that it'll use the secret set in
WEBWORK_SECRET env var if available. This way we don't have to customize
the webwork2.mojolicious.yml for each of our deployments.