EDIT customize error page, root url redirect
Can now customize the support email address that's shown on the default
error page with the new env var WEBWORK_SUPPORT_EMAIL

The support email will default to lthub since that address probably
won't go away anytime soon.

The root url shows a generic greeting page because the actual app lives
under the /webwork2 path. There's an option in webwork2.mojolicious.yml
that lets us redirect the root url to /webwork2, so users don't see the
generic greeting page. I've enabled this option. This redirect was in
the previous version of Webwork 2.17 but it was in Apache config which
no longer exists in 2.18.