FIX designers can't edit problem files
I originally patched the ContentGenerator for the problem editor so that
editing problem sets checks for "modify_problem_sets" permission instead
of the original "modify_student_data" permission (we don't want
designers modifying student data). This patch was lost when the
permission check was moved to the PGProblemEditor template upstream,
which also restored it to checking "modify_student_data".

Patching the template fixes this.