FIX upload file size limit too low
When the file size exceeds the limit, Mojolicious resets the connection.
Because this reset doesn't happen instantly, I've found that you can
exceed the limit by a little bit, if your file can make it onto the
server before the reset, it'll proceed as normal.

The default is 16MiB, but I've sucessfully uploaded a 18.8MiB file
during testing. A 80.9MiB file didn't make it.

I've added an env var so we can adjust the limit upward if needed, set
MAX_REQUEST_SIZE to the size we want to allow in bytes. I've defaulted
this to 1 GiB for now cause I can't seem to find what we used to allow.