FIX lti1p3 worker oom restarts due to memory leak
Normally, in garbage collected languages, you would expect that objects
which are unreachable gets garbage collected. Unfortunately, it seems
that Perl doesn't use this type of tracing garbage collection. Perl uses
what seems to be a basic reference counting gc that's not very robust,
so there's a ton of orphaned references that doesn't get collected.

As it seems that every db query leaks some objects due to this, we're
unlikely to fix this directly. Restarting seems to be the only feasible
way to sidestep this leak.

So for run_all_jobs.pl script which the worker_lti1p3 container runs,
I've split off the actual work into a separate thread. This thread gets
completes, exits, and then get recreated after completing 10 jobs. This
in effect gets us a "restart". To do this, we're no longer using
TheSchwartz's work(), instead using work_once() for more manual control
over processing the job queue.

The Mojolicious framework in the main app uses threads to fake async, so
I think it's avoided this particular pitfall due to that (hopefully).