FIX lti worker memory leak, replace thread w fork
The previous fix to the lti1p3 worker still resulted in a slowly
increasing memory consumption. Even explicitly undef the thread variable
to force cleanup of done threads didn't completely solve this. I notice
a slow uptick in memory usage on the order of megabytes in longer
testing. I don't know why, but ran across a recommendation in the perl
docs to not use perl's interpreter thread implementation, so maybe it
was a mistake to use threads in the first place.

So the immediate alternative is to replace thread with fork. Honestly, I
couldn't rule out a remaining memory leak completely during testing, but
at least it's a lot less obvious than compared to the thread
implementation. There might still be some kilobytes being unrecoverable
on each cycle with the fork implementation, this should hopefully be
neglible enough in prod even with long uptimes (assuming we reboot
Webwork at least once a year for updates).

To further reduce the effects of the possible leak on each fork, I've
extended the number of jobs processed before starting a new fork from 10
to 100.

I was under the impression that child processes would get the same
signal as parent processes, but that didn't happen during testing. I had
to explicitly pass on the same signal to the child in order to ensure
that the child process obeys the same SIGQUIT that the main Webwork app
does.