FIX speculative fix on empty lti launch failures
Users complain LTI login isn't working, Webwork shows an error page
indicating a POST request was sent with no headers and no body. One user
was using Chrome on MacOS. We were unable to replicate this.

So this is a completely speculative fix, the thinking is that we're
triggering CORS protection somehow. Maybe the empty request is
indicative of CORS preflight? Anyways, I tried to implement a response
to the CORS preflight for the LTI endpoints, so hopefully they'll be
allowed through.

There's a lot of reasons to think CORS isn't the issue, the main one
being that the LTI spec specifically says this has to be a form POST
request to avoid all these issues, and as far as I can see, Canvas does
render a regular html form for this. And form submits are supposed to be
exempt from CORS. But there might be JavaScript involved in the chain,
so perhaps on some systems it gets treated as an XHR request.

Added a config to localOverrides.conf, $lti_advantage{allow_origins}
specifies the list of domains that we can receive CORS LTI launches
from. For our purposes, this should just be Canvas. Note that each
origin must match exactly what you expect Canvas to be sending as the
Origin header. Canvas sends the Origin header on LTI launches, so I was
able to confirm that Canvas does not include a trailing / at the end of
their Origin.