FIX Canvas no longer recognize late assignments
During grade sync, the student's submit time is also sent to Canvas, so
Canvas can properly mark a student's assignment submission as late in
the Canvas gradebook. This seems to have broken now though, with Canvas
only marking some assignments as late.

The cause seems to be a change on the Canvas side (commit 818e211 in the
canvas-lms repo about 10 months ago) where only grade syncs with
activityProgress set to "Completed" or "Submitted" are treated as an
actual score request. Most of the time, Webwork sends grade syncs with
activityProgress set to "InProgress", with "Submitted" sent if the
student achieved a full score and "Completed" sent if the grade sync was
sent after the hmwk set's due date. I think the thinking behind this
decision was originally that a student will probably retry what they got
wrong until they got it right, so it seemed better to mark grades as
"InProgress" before they got it right. However, this is why Canvas is no
longer processing them properly.

The fix is to default the activityProgress to "Submitted".

Related to this, the submit time was being sent as a Canvas custom
extension parameter "https://canvas.instructure.com/lti/submission"
because the LTI AGS spec at the time didn't have a way to set student
submit time. At some point along the way, the LTI AGS spec added support
for submission.submittedAt param for this purpose. Canvas implemented
support for this last year, so we'll switch to using
submission.submittedAt too.