FIX can't create Canvas Test Student user anymore
Canvas variable substitution used to return empty string if the
substitution cannot be made. But I think at some point, it started
returning the variable name instead, e.g.: if a user has no PUID, it
returns $Canvas.user.sisIntegrationId

This caused the LTI NRPS sync to fail on the Test Student, because the
Test Student doesn't have a PUID and the $ in the variable name isn't
allowed as a Webwork user identifier.

Fixed this by checking if the user identifier or student number starts
with the Canvas variable substitution prefix $Canvas. If it's a Canvas
variable substitution, then it gets set to the default value. For the
user identifier, this because the LTI user ID and for the student
number, it's empty string.